Experiment to see if changing baseurl fixes issue
Site does not render on github pages.  Trying to understand why.  Given the issues with the way this page appeared, I think the problem has to do with the way files are being accessed from the repo (or not accessed).  I changed the base URL to https://johndeanossowski.github.io/PortfolioSite/ to see if this corrected the issue.